refactor: put birds data to separate files
