+++ src/common/data/chickenlike/index.js
@@ -1,0 +1,3 @@
+import chickenlike from './chickenlike';
+
+export default chickenlike;
