+++ src/common/data/data.js
@@ -1,258 +1,17 @@
-const owls = [
-    {
-        id: 1,
-        name: 'Филин',
-        latinName: 'bubo-bubo',
-        description: 'Филин ведёт оседлый образ жизни, придерживается своей территории, активен в сумерки и ночью. Обладает отличным слухом и зрением. Питается мелкими грызунами и птицами. Филины живут по одиночке в дуплах и гнёздах крупных птиц. Приносят потомство раз в год',
-    },
-    {
-        id: 2,
-        name: 'Сипуха',
-        latinName: 'tyto-alba',
-        description: 'Сипухи - хищные осёдлые птицы. Сипухи селятся в дуплах, на чердаках заброшенных домов. Спят днём, охотятся ночью. Сипухи предпочитают искать добычу в лугах, полях, сельскохозяйственных угодьях.',
-    },
-    {
-        id: 3,
-        name: 'Мохноногий сыч',
-        latinName: 'aegolius-funereus',
-        description: 'Мохноногий сыч - небольшая сова, размером с голубя. Своё название полечил из-за оперения: оно доходит до самых когтей. Ведёт оседлый образ жизни, гнездится в лесах, обычно в хвойных. Охотится на мелких лесных зверьков.',
-    },
-    {
-        id: 4,
-        name: 'Серая неясыть',
-        latinName: 'strix-aluco',
-        description: 'Серая неясыть можно встретить в старых лесах различного типа. глухих парках, деревнях. Птицы размером чуть больше вороны. Самцы издают узнаваемый звук "у-хуу-уу", у самок другой крик - короткое "ку-вит". Серая неясыть может вращать головой на 270 градусов.',
-    },
-    {
-        id: 5,
-        name: 'Ушастая сова',
-        latinName: 'asio-otus',
-        description: 'Ушастая сова предпочитает гнездится в местах, освоенных человеком: кладбищенских рощах, выгонах, лесозащитных полосах и т.д. Днём прячется в кронах деревьев, активна ночью. Птица выделяется большими перьевыми "ушками" на голове. Самый распространённый вид сов в Беларуси',
-    },
-    {
-        id: 6,
-        name: 'Воробьиный сычик',
-        latinName: 'glaucidium-passerinum',
-        description: 'Воробьиный сычик получил название из-за своих размеров. Самка весит 63-103 г., длина тела - 17-19см. Это самая маленькая сова Беларуси. Сычик гнездится в хвойных лесах, питается в основным мышевидными грызунами и бурозубками. Сычик делает запасы: иногда в кладовых находят до 40 мышей-полёвок.',
-    }
-]
-
-const dayPredators = [
-    {
-        id: 1,
-        name: 'Змееяд',
-        latinName: 'circaetus-gallicus',
-        description: 'Змееяд питается почти исключительно змеями. Пуглив и недоверчив по отношению к человеку, гнездится в удалённых местах. Змееяд предпочитает вить гнёзда на самых макушках деревьев, тщательно скрывая их снизу и с боков. В поисках добычи может отлетать от гнезда на расстояние до 13 км. ',
-    },
-    {
-        id: 2,
-        name: 'Канюк',
-        latinName: 'buteo-buteo',
-        description: 'Птица средней величины: длина тела до 57 см, размах крыльев до 128 см. Канюка легче заметить во время охоты: он сидит на ветке дерева на опушке и высматривает грызунов. Птицы селятся одиночными парами. Канюк в Беларуси широко распространён.',
-    },
-    {
-        id: 3,
-        name: 'Скопа',
-        latinName: 'pandion-haliaetus',
-        description: 'Скопа обитает возле водоёмов, богатых рыбой. Питается исключительно рыбой, иногда может добыть мышей или лягушек. В Беларуси чаще встречается возле Западной Двины. Предпочитает гнездиться в болотах, подальше от человека. У одной пары птиц может быть до 5 гнёзд.',
-    },
-    {
-        id: 4,
-        name: 'Лунь болотный',
-        latinName: 'circus-aeruginosus',
-        description: 'Птица обитает возле болот и зарастающих озёр. Гнёзда утраивает на кочке, отмели или в зарослях рогоза или камыша. Во время охоты плавно скользит в воздухе, увидев добычу, поднимает вверх крылья, опускает лыпа и падает вертикально вниз.',
-    },
-    {
-        id: 5,
-        name: 'Орлан-белохвост',
-        latinName: 'haliaeetus-albicilla',
-        description: 'Самая крупная хищная птица Беларуси: самки крупнее, весят до 6,5 кг, длина тела - 85-98 см. Обитает возле водоёмов. Гнёзда очень большие, строятся из толстых веток и сучьев. Орлан питается птицами (утка, гусь, лысуха), мелкопитающими, рыбой, а также падалью.',
-    },
-    {
-        id: 6,
-        name: 'Ястреб-тетеревятник',
-        latinName: 'accipiter-gentilis',
-        description: 'Тетеревятник - сильная и агрессивная птица. Он охотится на птиц, размерами до тетерева и домашних кур. Может съесть других хищных птиц - луней, канюков. Пара тетеривятников живёт на гнездовом участке круглый год. Селятся обычно в глухих лесах, особенно заболоченных.',
-    },
-];
-
-const sparrowlike = [
-    {
-        id: 1,
-        name: 'Жаворонок полевой',
-        latinName: 'alauda-arvensis',
-        description: 'Жаворонок чуть крупнее воробья. Весной прилетает одной из первых. Селится на открытых травянистых пространсвах, часто встречается на обрабатываемых полях. Хорошо известен благодаря своей звонкой и заливистой песне. Питается насекомыми, пауками и семенами растений.',
-    },
-    {
-        id: 2,
-        name: 'Канареечный вьюрок',
-        latinName: 'serinus-serinus',
-        description: 'Размер вьюрка - всего 11-12 см. Название птица получила за ярко-жёлтую окраску. Для вьюрка характерен очень короткий клюв. Вьюрок селится на ухоженных территориях с кустами и деревьями: в садах, живых изгородях, парках.',
-    },
-    {
-        id: 3,
-        name: 'Грач',
-        latinName: 'corvus-frugilegus',
-        description: 'У грача полносью чёрное оперение. Птица размером с ворону. Грачи предпочитают селиться вблизи человека: на полях, в населённых пунктах. Гнездятся колониями. Пение грача - хриплое "каар".',
-    },
-    {
-        id: 4,
-        name: 'Дрозд певчий',
-        latinName: 'turdus-philomelos',
-        description: 'Песня дрозда очень разнообразна: обычно состоит из флейтовых звуков, посвистов. Птица поёт долго, повторяя части 2-3 раза. Дрозд скрытен, но при приближении человека к гнезду издаёт серию трескучих звуков.',
-    },
-    {
-        id: 5,
-        name: 'Дубонос',
-        latinName: 'coccothraustes-coccothraustes',
-        description: 'Дубонос держится скрытно, в верху кроны деревьев. Характер нордический, выдержанный. Питается семенами граба, вишни, черёмухи, почками деревьев. Птенцов выкармливает насекомыми.',
-    },
-    {
-        id: 6,
-        name: 'Галка',
-        latinName: 'coloeus-monedula',
-        description: 'Галки похожи на ворон, но мельче размерами. Широко распространены, особенно в окультуренных местах. Криклива, голос напоминает "кау" или "чак". Держится стаями.',
-    },
-];
-
-const chickenlike = [
-    {
-        id: 1,
-        name: 'Глухарь',
-        latinName: 'tetrao-urogallus',
-        description: 'Глухарь селиться в старых глухих лесах. Птицы фактические не образуют пар, о потомстве заботиться только самка. Для спаривания глухари встречаются на "токовищах". Крик глухаря похож на щёлкание по пустому сосуду.',
-    },
-    {
-        id: 2,
-        name: 'Фазан',
-        latinName: 'phasianus-colchicus',
-        description: 'На фото - самец фазана. Самки имеют равномерное серо-бурое оперение. Оседлая птица, почти всё время проводит на земле, иногда взлетая на ветки деревьев. Птицы склёвывают ягоды, семена, входы растений. Ценный охотничий вид.',
-    },
-    {
-        id: 3,
-        name: 'Серая куропатка',
-        latinName: 'perdix-perdix',
-        description: 'Куропатки большую часть времени проводят на земле или в кустарниках. Птица быстро бегает, в случае опасности резко и шумно взлетает, чем обескуращивает хищника. Птицы ведут оседлый образ жизни, живут стаями.',
-    },
-    {
-        id: 4,
-        name: 'Перепел',
-        latinName: 'coturnix-coturnix',
-        description: 'Перепел ведёт исключительно наземный образ жизни, на деревья не садится. Кормится растительной пищей, летом употребляет пауков. Это единственная перелётная птица семейства курообразных: в Беларусь перепел прилетает в середине мая, в конце августа начинается отлёт.',
-    },
-    {
-        id: 5,
-        name: 'Рябчик',
-        latinName: 'tetrastes-bonasia',
-        description: 'Название "Рябчик" полностью характеризует окраску птицы: её тело покрыто белыми, бурыми, серыми и коричневыми пестринами. Рябчика сложно заметить на ветке дерева. Эта птица оседла, лишь незначительно мигрирует в пределах леса. Рябчик ведёт себя тихо, но её посвист можно услышать в брачный период.',
-    },
-    {
-        id: 6,
-        name: 'Тетерев',
-        latinName: 'lyrurus-tetrix',
-        description: 'Тетерев любит селиться в молодых лесах, на местах гарей и вырубок. Это оседлая птица, не образует пар. Спаривание происходит на току, после чего самцы удаляются в глухие места, а самки - откладывать яйца.',
-    },
-];
-
-const sparrowlike1 = [
-    {
-        id: 1,
-        name: 'Зяблик',
-        latinName: 'fringilla-coelebs',
-        description: 'Зяблика можно встретить во различных лесах, городских парках, садах. Птица может поедать любые семена трав, цветки и насекомых. У зяблика нет выраженной специализации, поэтому он широко распространён. Голос зяблика - громкая и быстрая трель.',
-    },
-    {
-        id: 2,
-        name: 'Иволга',
-        latinName: 'oriolus-oriolus',
-        description: 'Тело и голова иволги ярко-жёлтые. Птицу можно встретить в лесах, садах, парках, она избегает больших и глухих массивов. Гнездо иволги особенное - птица подвешивает его между ветвями наподобие гамака. Питается насекомыми и их личинками.',
-    },
-    {
-        id: 3,
-        name: 'Клёст-еловик',
-        latinName: 'loxia-curvirostra',
-        description: 'Птица получила название за своё пение: быстрые позывки "кле-кле-кле". Характерная черта: концы надклювья и подклювья крестообразно загнуты вверх и вниз соответственно. Питаются в основном семенами ели, реже - сосны и лиственницы.',
-    },
-    {
-        id: 4,
-        name: 'Желтоголовый королёк',
-        latinName: 'regulus-regulus',
-        description: 'Королёк - одна из самых маленьких птиц Беларуси. Самец весит 5-7 г., длина тела - 8-10,5 см. У взрослых птиц перья на голове ярко-жёлтые. Птица населяет старые ельники, осенью встречается в садах и парках. Королёк цепляется лапками за ветки и принимает разные позы.',
-    },
-    {
-        id: 5,
-        name: 'Овсянка обыкновенная',
-        latinName: 'emberiza-citrinella',
-        description: 'Овсянку можно отличить по ярко-жёлтому оперению головы и груди. Птицы живут парами. Самец овсянки поёт обычно на вершине дерева, показывая, что гнездовой участок охраняется. Гнездод птиц располагается в ямке на земле.',
-    },
-    {
-        id: 6,
-        name: 'Свиристель',
-        latinName: 'bombycilla-garrulus',
-        description: 'Свиристели гнездятся в Беларуси с октября по апрель. Свое присутствие птицы выдают нежными высокими позывками «свирри-цирр», определившими русское название вида. А вот белорусское название этой птицы - амялушка, - обязано ее пристрастием к ягодам омелы.',
-    },
-]
-
-const differentBirds = [
-    {
-        id: 1,
-        name: 'Щегол',
-        latinName: 'carduelis-carduelis',
-        description: 'Щеглы выделяются красным оперением "лица". Летом птица гнездится парами, осенью и зимой держится в стаях. Щеглы предпочитают лесисту местность, где есть открытые места: луга, поляны. Во время строительства гнезда птицы непрерывно поют.',
-    },
-    {
-        id: 2,
-        name: 'Скворец',
-        latinName: 'sturnus-vulgaris',
-        description: 'Оперение скворца чёрное с фиолетовым отливом. Выделяется жёлтый клюв. Пение скворца оцень разнообразно: он копирует трели у других птиц, а также другие звуки, например, скрип двери или рычание мотора. Во время пения самец трясёт крыльями.',
-    },
-    {
-        id: 3,
-        name: 'Сойка',
-        latinName: 'garrulus-glandarius',
-        description: 'У сойки характерная внешность: на спине, животе, груди и голове светло-бурое оперение. Птица громко поёт, издаёт хриплый крик "крээч". Также сойка хорошо копирует пение других птиц, особенно дневных хищников (коршуна, канюка). Птица питается семенами, овощами, желудями. На зиму делает запасы.',
-    },
-    {
-        id: 4,
-        name: 'Чибис',
-        latinName: 'vanellus-vanellus',
-        description: 'Крик чибиса - "чьии-вы". Птица часто и громко подаёт голос. Чибис предпочитает увлажнённые места - луга, болота. Гнездо птица располагает не маскируя, на кочке или гребне земли. Птенцы покидают гнездо в первый день, но ещё месяц находятся под опекой родителей.',
-    },
-    {
-        id: 5,
-        name: 'Бекас',
-        latinName: 'gallinago-gallinago',
-        description: 'Бекаса отличает очень длинный тонкий клюв и короткие ноги. Обитает в заболоченных поймах рек, на болотах, влажных лугах. Ведёт ночной и сумеречный образ жизни. Бекас летит очень быстро и зигзагами, что спасает его от хищников.',
-    },
-    {
-        id: 6,
-        name: 'Дятел пёстрый',
-        latinName: 'dendrocopos-major',
-        description: 'В окрасе дятла преобладают чёрные и белые тона. Птица известна благодаря способу добычи пищи: дятел прыжками перемещается по стволу дерева, ища насекомых. Когда находит, он разбивает кору клювом и достаёт их языком.',
-    },
-];
-
-const imgURL = 'https://github.com/RadyukAG/songbird/blob/songbird/src/img/Birds/';
-const audioURL = 'https://github.com/RadyukAG/songbird/tree/songbird/src/sounds/birds';
-
-function addURLs(arr) {
-  return arr.map((obj) => {
-    const newObj = obj;
-    newObj.image = `${imgURL}${newObj.latinName}.jpg?raw=true`;
-    newObj.audio = `${audioURL}${newObj.latinName}.mp3`;
-    return obj;
-  });
-}
+import owls from './owls';
+import dayPredators from './dayPredators';
+import sparrowlike from './sparrowLike';
+import chickenlike from './chickenlike';
+import sparrowlike1 from './sparrowlike1';
+import differentBirds from './differentBirds';
 
 const data = {
-    'Совиные': addURLs(owls),
-    'Дневные хищники': addURLs(dayPredators),
-    'Воробьиные (часть 1)': addURLs(sparrowlike),
-    'Курообразные': addURLs(chickenlike),
-    'Воробьиные (часть 2)': addURLs(sparrowlike1),
-    'Разные': addURLs(differentBirds),
+    'Совиные': owls,
+    'Дневные хищники': dayPredators,
+    'Воробьиные (часть 1)': sparrowlike,
+    'Курообразные': chickenlike,
+    'Воробьиные (часть 2)': sparrowlike1,
+    'Разные': differentBirds,
 }
-
-
 
 export default data;
