+++ src/common/data/dayPredators/index.js
@@ -1,0 +1,3 @@
+import dayPredators from './dayPredators';
+
+export default dayPredators;
