+++ src/common/data/differentBirds/index.js
@@ -1,0 +1,3 @@
+import differentBirds from './differentBirds';
+
+export default differentBirds;
