+++ src/common/data/owls/index.js
@@ -1,0 +1,3 @@
+import owls from './owls';
+
+export default owls;
