+++ src/common/data/sparrowLike/index.js
@@ -1,0 +1,3 @@
+import sparrowlike from './sparrowLike';
+
+export default sparrowlike;
