+++ src/common/data/sparrowlike1/index.js
@@ -1,0 +1,3 @@
+import sparrowlike1 from './sparrowlike1';
+
+export default sparrowlike1;
